FastAPI/requirements.txt
PyYAML~=6.0
pydantic~=1.10.6
uvicorn~=0.21.1
fastapi~=0.94.1
starlette~=0.26.1
requests~=2.28.2